correcting positions
