+++ components/BookmarkIcon.js
@@ -1,0 +1,13 @@
+import {Icon} from "@iconify/react";
+import styled from "styled-components";
+
+
+const StyledIcon = styled(Icon)`
+width: 40px;
+height: 50px;
+position: absolute;
+top: 6.5rem;
+right: 1.5rem;
+`;
+
+export default StyledIcon;
