+++ pages/bestellung/verzug.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import Link from "next/link";
-import {Icon} from "@iconify/react";
+import StyledIcon from "../../components/BookmarkIcon";
 import Breadcrumb from "../../components/Breadcrumb";
 import ZurueckButton from "../../components/Zurueckbutton";
 import StyledBody from "../../components/StyledBody";
@@ -41,12 +41,4 @@ const StyledArticle = styled.article`
 
 const StyledLink = styled(Link)`
   text-decoration: none;
-`;
-
-const StyledIcon = styled(Icon)`
-  width: 40px;
-  height: 50px;
-  position: absolute;
-  top: 6.5rem;
-  right: 2rem;
 `;
